requirements.txt
# pip install -r requirements.txt
PyOpenGL 
PyOpenGL_accelerate
valideer
Pillow
numpy